Worked on main file
Got a majority of the way through main. Only have to finish the handler for when the user is inside the cart. I believe it should all work correctly, but I was not able to test anything due to the User class not being completed. 

I also changed the constructors inside the User class so they align with the design document